renamed some files for consistency
