+++ src/client/sections/video-section.tsx
@@ -1,8 +1,8 @@
 import React, { useState} from "react";
-import DeviceSelector from "../components/deviceselector";
-import { JWTCapability } from "../components/jwtcapability";
+import DeviceSelector from "../components/device-selector";
+import { JWTCapability } from "../components/jwt-capability";
 
-export function HealthSection() {
+export function VideoSection() {
   const [device, setDevice] = useState();
 
   return (
@@ -18,7 +18,14 @@ export function HealthSection() {
         <div
           className="flex flex-1 items-center justify-center rounded-lg border border-dashed shadow-sm"
         >
-          {device && <JWTCapability device={device} capability={"@transitive-robotics/health-monitoring"} delimiters={"undefined"}/>}
+          {device && <JWTCapability 
+            device={device}
+            capability={"@transitive-robotics/webrtc-video"}
+            count="1"
+            quantizer="25"
+            timeout="1800"
+            type="videotestsrc"
+          />}  
         </div>
       </main>
     </div>
